+++ internal_testsuites/golang/testsuite/startosis_upload_files_test/startosis_upload_files_test.go
@@ -62,7 +62,7 @@ func TestStartosis(t *testing.T) {
 
 	outputStream, _, err := enclaveCtx.ExecuteKurtosisScript(ctx, startosisScript, defaultDryRun)
 	require.NoError(t, err, "Unexpected error executing startosis script")
-	interpretationError, validationErrors, executionError, instructions := test_helpers.ReadStreamContentUntilClosed(outputStream)
+	scriptOutput, _, interpretationError, validationErrors, executionError := test_helpers.ReadStreamContentUntilClosed(outputStream)
 
 	expectedScriptOutput := `Adding service example-datastore-server-1.
 Uploaded [a-f0-9-]{36}
@@ -71,7 +71,7 @@ Uploaded [a-f0-9-]{36}
 	require.Nil(t, interpretationError, "Unexpected interpretation error. This test requires you to be online for the upload_file command to run")
 	require.Empty(t, validationErrors, "Unexpected validation error")
 	require.Nil(t, executionError, "Unexpected execution error")
-	require.Regexp(t, expectedScriptOutput, test_helpers.GenerateScriptOutput(instructions))
+	require.Regexp(t, expectedScriptOutput, scriptOutput)
 	logrus.Infof("Successfully ran Startosis script")
 
 	// Check that the service added by the script is functional